+++ frontend/src/LocationCard/data.js
